Fixed Movement
Now you can move without having the game close. The game works for the most part. The major flaw at the moment is that after the goose goes over the wall it doesn't reappear. The solution for this is to either have the goose completely skip over the wall or have the wall reappear after the goose hops over it.